shell: Fix signed-unsigned comparisons

+++ src/shell-app.c
@@ -699,7 +699,7 @@ static int
 shell_app_get_last_user_time (ShellApp *app)
 {
   GSList *iter;
-  int last_user_time;
+  guint32 last_user_time;
 
   last_user_time = 0;
 
@@ -709,7 +709,7 @@ shell_app_get_last_user_time (ShellApp *app)
         last_user_time = MAX (last_user_time, meta_window_get_user_time (iter->data));
     }
 
-  return last_user_time;
+  return (int)last_user_time;
 }
 
 /**
